v1.21
Updated the game UI to show times and scores as NA if there isnt a saved score and time for the number of pieces created.
Added options button to game screen so player has access to the options page instead of only being able to make changes at start or end of the game.